Rename file to utils and add test
