+++ modelstore/metadata/dataset/features.py
@@ -21,7 +21,7 @@ from modelstore.metadata.dataset.types import (
     is_numpy_array,
     is_pandas_dataframe
 )
-from modelstore.metadata.utils.dicts import exclude_field
+from modelstore.metadata.utils.utils import exclude_field
 from modelstore.utils.log import logger
