+++ modelstore/metadata/model/model_type.py
@@ -18,7 +18,7 @@ from dataclasses import dataclass, field
 from dataclasses_json import dataclass_json
 from dataclasses_json.cfg import config
 
-from modelstore.metadata.utils.dicts import exclude_field
+from modelstore.metadata.utils.utils import exclude_field
 
 _MODEL_TYPE_FILE = "model-info.json"
