+++ tests/models/test_sklearn.py
@@ -15,7 +15,6 @@ import json
 import os
 from functools import partial
 
-import numpy as np
 import pandas as pd
 import pytest
 from sklearn.compose import ColumnTransformer
